==> 4a-ephys-analysis.Rmd <==
---
title: "4-ephys-analysis"
author: "Keon Arbabi"
date: "19/01/2022"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# load libraries 
```{r}
suppressPackageStartupMessages({
  library(tidyverse)
  library(tidytext)
  library(magrittr)
  library(here)
  library(data.table)
  library(lme4)
  library(lmerTest)
  library(MuMIn)
  library(GGally)
  library(cowplot)
  library(gridExtra)
  library(ggpubr)
  library(mice)
  library(stringr)
  library(svglite)
  library(performance)
  library(broom.mixed)
  library(dotwhisker)
})
```

# load data
```{r}
### can we get more updated cell metadata from here?? https://portal.brain-map.org/explore/classes/multimodal-characterization

# load summarized ephys features 
ephys_h = read.csv(file = here("data","processed_ephys","human_ephys_features_berg.csv"))
# ephys_h = read.csv(url('https://raw.githubusercontent.com/AllenInstitute/patchseq_human_L23/master/data/human_mouse_ephys_all_0127.csv')) %>% 
ephys_h = read_csv('/external/rprshnas01/netdata_kcni/stlab/Public/AIBS_patchseq_2020/human/ephys/processed/ephys_features_berg_w_QC.csv')
  # rename(ephys_session_id = cell_specimen_id)
#   rename(ephys_session_id = specimen_id)

# load microglia scores 
qcMetrics_h = read.csv(file = here("output","human_qcMetrics.csv")) %>% dplyr::select(c(sample_id, Microglia))
# load metadata
metadata_h = fread(file = here("data","patchseq","20200625_patchseq_metadata_human.csv"), data.table = FALSE) %>% 
  dplyr::rename(sample_id = transcriptomics_sample_id, cluster_label = corresponding_AIT2.3.1_alias) %>%
  mutate(cluster_label = factor(cluster_label,
                                levels = c("Exc L2 LAMP5 LTK","Exc L2-4 LINC00507 GLP2R","Exc L2-3 LINC00507 FREM3","Exc L3-4 RORB CARM1P1","Exc L3-5 RORB COL22A1"),
                                labels = c("LTK","GLP2R","FREM3","CARM1P1","COL22A1"))) %>%
  dplyr::select(c(cell_specimen_id, ephys_session_id, sample_id, donor_id, cell_soma_normalized_depth, cluster_label, depth_from_pia_um, cell_specimen_name)) %>%
  left_join(., qcMetrics_h, by = "sample_id") %>%
  left_join(., ephys_h, by = "cell_specimen_id")

# load summarized ephys features 

# patch-seq mouse
qcMetrics_m = read.csv(file = here("output","mouse_qcMetrics.csv"))

metadata_m = fread(file = here("data","patchseq","20200711_patchseq_metadata_mouse.csv"), data.table = FALSE) %>%
  dplyr::rename(sample_id = transcriptomics_sample_id)
metadata_m = merge(metadata_m, qcMetrics_m %>% dplyr::select(c("sample_id","major_type","contam_type","quality_score","Microglia")), by = "sample_id")
metadata_m = metadata_m[colSums(!is.na(metadata_m))>0]
rownames(metadata_m) = metadata_m$sample_id

metadata_m = metadata_m %>% dplyr::rename(corresponding_AIT2.3.1_alias = `T-type Label`) # do this cause we iniitally wrote the code with the older metadata sheet and fields

metadata_m = metadata_m %>% mutate(cluster_label = str_extract(corresponding_AIT2.3.1_alias, "\\w+")) 
metadata_m = metadata_m %>% separate(corresponding_AIT2.3.1_alias, into = c("subclass", "second")) %>% 
  unite(col = "corresponding_AIT2.3.1_alias", subclass, second, sep = " ")
leaveout = metadata_m %>% dplyr::count(corresponding_AIT2.3.1_alias, sort = TRUE) %>% filter(n < 20) %>% pull(corresponding_AIT2.3.1_alias) %>% unique()
leaveout = c(leaveout,"",NA, " NA")
metadata_m = metadata_m %>% filter(!corresponding_AIT2.3.1_alias %in% leaveout) 

metadata_m$cluster_label = factor(metadata_m$cluster_label, 
                                  levels = c("Lamp5","Pvalb","Serpinf1","Sncg","Sst","Vip"),
                                  labels = c("Lamp5","Pvalb","Serpinf1","Sncg","Sst","Vip"),
)

metadata_m = metadata_m %>% separate(cell_specimen_name, into = c("first", "slice_num_first", "slice_num_second", "cell_num"), remove = F, sep = '\\.') 

metadata_m = metadata_m %>% mutate(slice_name = paste(slice_num_first, slice_num_second, sep = '.'), 
                                   slice_id = paste(donor_id, slice_name, sep = '.'))

ephys_m = read.csv('/external/rprshnas01/netdata_kcni/stlab/Public/AIBS_patchseq_2020/mouse/ephys/processed/ephys_features_gouwens_w_QC.csv')
metadata_m = metadata_m %>% 
  left_join(., ephys_m, by = "cell_specimen_id")

metadata_m = metadata_m %>% filter(!qc_recording_date == '') %>%
  mutate(qc_recording_date = as.POSIXct(qc_recording_date)) %>%
  group_by(donor_id) %>% 
  arrange(qc_recording_date) %>%
  mutate(first_cell_time = first(qc_recording_date), 
         rel_recording_time = as.numeric(difftime(qc_recording_date, first_cell_time, units = "hours")))

```

# transform ephys properties as needed
```{r}

metadata_h %<>% mutate(
                        input_resistance = log10(input_resistance), 
                        rheobase_i = log10(rheobase_i), 
                        width = log10(rheo_width * 1000),
                        apamp = rheo_peak_v - rheo_threshold_v, 
                        ahpamp = rheo_threshold_v - rheo_fast_trough_v,
) %>%
  rename(trough_v = rheo_trough_v,
         threshold_v = rheo_threshold_v)

metadata_m %<>% mutate(
                        input_resistance = log10(input_resistance), 
                        rheobase_i = log10(rheobase_i), 
                        width = log10(rheo_width * 1000),
                        apamp = rheo_peak_v - rheo_threshold_v, 
                        ahpamp = rheo_threshold_v - rheo_fast_trough_v,
) %>%
  rename(trough_v = rheo_trough_v,
         threshold_v = rheo_threshold_v)

model_features_list = c('input_resistance', 'rheobase_i', 'width', 'apamp', 'ahpamp', 'threshold_v', 'trough_v', 'fi_fit_slope', 'v_baseline', 'sag', 'adapt')

```

# generate some scatter plots to illustrate what might be happening with contamination
```{r}
# subset to samples with cell soma normalized depth and microglia scores 
# metadata_h = metadata_h[!is.na(metadata_h$cell_soma_normalized_depth),]
metadata_h = metadata_h[!is.na(metadata_h$Microglia),]
metadata_h$color = 'black'
metadata_h$size = 1
# low microglia cell example: 732048771	SM-GE8YK_S028_E1-50	H18.03.314	ephys_session_id sample_id donor_id # sub-731978186_ses-732048771_icephys.nwb, #355 um deep
# high microglia cell example: 732091715	SM-GE8YK_S031_E1-50	H18.03.314 ephys_session_id sample_id donor_id # sub-731978186_ses-732091715_icephys.nwb #331 um deep
# nwb datasets are on the scc here: /external/rprshnas01/netdata_kcni/stlab/Public/AIBS_patchseq_2020/human/ephys/000023/
metadata_h[metadata_h$ephys_session_id == 732048771, 'color'] = '#357EBDFF'
metadata_h[metadata_h$ephys_session_id == 732048771, 'size'] = 5
metadata_h[metadata_h$ephys_session_id == 732091715, 'color'] = '#D43F3AFF'
metadata_h[metadata_h$ephys_session_id == 732091715, 'size'] = 5


# subset to samples with cell soma normalized depth and microglia scores 
# metadata_h = metadata_h[!is.na(metadata_h$cell_soma_normalized_depth),]
metadata_m = metadata_m[!is.na(metadata_m$Microglia),]
metadata_m$color = 'black'
metadata_m$size = 1
# low microglia cell example: 892441714 sub-885375443_ses-892441714_icephys.nwb
# high microglia cell example: 892424119	sub-885375443_ses-892424119_icephys.nwb
# nwb datasets are on the scc here: /external/rprshnas01/netdata_kcni/stlab/Public/AIBS_patchseq_2020/human/ephys/000023/
metadata_m[metadata_m$ephys_session_id == 892441714, 'color'] = '#357EBDFF'
metadata_m[metadata_m$ephys_session_id == 892441714, 'size'] = 5
metadata_m[metadata_m$ephys_session_id == 892424119, 'color'] = '#D43F3AFF'
metadata_m[metadata_m$ephys_session_id == 892424119, 'size'] = 5

metadata_m %>% filter(corresponding_AIT2.3.1_alias == 'Pvalb Sema3e', donor_id == 470303) %>% 
  group_by(donor_id) %>% mutate(cells_per_donor = n()) %>% ungroup() %>%
  select(Microglia, cell_soma_normalized_depth, input_resistance, trough_v, v_baseline, everything()) %>% 
  as.data.frame() %>% mutate(input_resistance = 10^input_resistance) %>% arrange(-cells_per_donor, donor_id, Microglia)

human_plot_rin = metadata_h %>% filter(cluster_label == 'FREM3') %>% 
  ggplot(aes(x = Microglia, y = 10^input_resistance, color = color, size = size)) + 
  geom_smooth(method = "lm", se = F) +
  geom_point(shape = 16) + 
  scale_color_identity() + 
  scale_size_identity() + 
  scale_y_log10() + 
  ylab('Input resistance (MΩ)') + 
  xlab('Microglia score') + 
  theme_classic() + 
  theme(axis.text = element_text(size = 12),
        axis.title = element_text(size = 16))  +
  stat_cor(method = "pearson", size = 4, label.x = 0, label.y = 2.5)

human_plot_ap_trough = metadata_h %>% filter(cluster_label == 'FREM3') %>% 
  ggplot(aes(x = Microglia, y = trough_v, color = color, size = size)) + 
  geom_smooth(method = "lm", se = F) +
  geom_point(shape = 16) + 
  scale_color_identity() + 
  scale_size_identity() + 
  ylab('AP trough (mV)') + 
  xlab('Microglia Score') + 
  theme_classic() +
  theme(axis.text = element_text(size = 12),
        axis.title = element_text(size = 16))  +
  stat_cor(method = "pearson", size = 4, label.x = 0, label.y = -30)

mouse_gouwens_plot_rin = metadata_m %>% filter(10^input_resistance > 50, 10^input_resistance < 300) %>% 
  filter(corresponding_AIT2.3.1_alias == 'Pvalb Sema3e') %>% 
  ggplot(aes(x = Microglia, y = 10^input_resistance, color = color, size = size)) + 
  geom_smooth(method = "lm", se = F) +
  geom_point(shape = 16, alpha = 1) + 
  scale_color_identity() + 
  scale_size_identity() + 
  scale_y_log10() + 
  # ylim(c(100, 650)) + 
  ylab('Input resistance (MΩ)') + 
  xlab('Microglia score') + 
  theme_classic() + 
  theme(axis.text = element_text(size = 12),
        axis.title = element_text(size = 16))  +
  stat_cor(method = "pearson", size = 4, label.x = 0, label.y = 2.5)

mouse_gouwens_plot_ap_trough = metadata_m %>%  
  filter(v_baseline < -50) %>% 
  filter(corresponding_AIT2.3.1_alias == 'Pvalb Sema3e') %>% 
  ggplot(aes(x = Microglia, y = v_baseline, color = color, size = size)) + 
  geom_smooth(method = "lm", se = F) +
  geom_point(shape = 16, alpha = 1) + 
  scale_color_identity() + 
  scale_size_identity() + 
  # ylim(c(100, 650)) + 
  ylab('RMP (mV)') +  
  xlab('Microglia score') + 
  theme_classic() + 
  theme(axis.text = element_text(size = 12),
        axis.title = element_text(size = 16))  +
  stat_cor(method = "pearson", size = 4, label.x = 0, label.y = -50)

ephys_example_plots = plot_grid(human_plot_rin, human_plot_ap_trough, mouse_gouwens_plot_rin, mouse_gouwens_plot_ap_trough, nrow = 2)
ephys_example_plots
ggsave(filename = "./output/figure_4a.svg", device = "svg", width = 7, height = 9, units = "in", dpi = 500, limitsize = F)
ggsave(filename = "./output/figure_4a.png", device = "png", width = 7, height = 9, units = "in", dpi = 500, limitsize = F)

metadata_m %>% 
  group_by(corresponding_AIT2.3.1_alias) %>% 
  summarize(corr_val_rmp = cor(Microglia, v_baseline, use = 'pairwise.complete.obs'),
            corr_val_rin = cor(Microglia, input_resistance, use = 'pairwise.complete.obs'),
            corr_val_trough_v = cor(Microglia, trough_v, use = 'pairwise.complete.obs'), 
            corr_val_fi_slope = cor(Microglia, fi_fit_slope, use = 'pairwise.complete.obs'),
            count = n(), 
            micro_sd = sd(Microglia)) %>% 
  filter(count > 100) %>%
  arrange(-corr_val_rmp) %>% 
  as.data.frame()

```

# linear modelling 
```{r}
# have discussion about what terms to use in each model - IMO we should use depth_from_pia_um as it's avail for most human samples but 
# no depth norm term for mice as it's not present for most mouse samples
form = scale(input_resistance) ~ 0 + (1|corresponding_AIT2.3.1_alias) + Microglia + (1|donor_id)
lmer_model = lmer(form, data = metadata_m, REML = FALSE)
form = scale(input_resistance) ~ 0 + (1|cluster_label) + Microglia + (1|donor_id) + scale(depth_from_pia_um)
lmer_model = lmer(form, data = metadata_h, REML = FALSE)
# view features lacking enough data 
n_count = sapply(metadata_h, function(x) sum(!is.na(x))) %>% as.data.frame() %>% rownames_to_column(var = "feature")
# remove
# metadata_h = metadata_h %>% dplyr::select(-c(slow_trough_v, slow_trough_t, adp_v, rheo_first_isi, adapt))
ephys_features = names(metadata_h)[7:ncol(metadata_h)]
meta_features = names(metadata_h)[(c(1:3, 4, 5:6))]
plot_lst = list()
# loop through ephys features 
# note slight difference in models below for mouse vs human - see explanation above
# the choice not to scale microglia is intentional as it makes it possible to compare terms between mouse and human more directly
# (they have different SDs, they're implicitly scaled by this)
keep_df = data.frame()
for(f in model_features_list){
  # prepare data 
  formula = paste0('scale(', f, ') ~ 1 + (1|cluster_label) + Microglia + (1|donor_id) + scale(depth_from_pia_um)')
  print(formula)
  tidy_df = tidy(lmer(formula, data = metadata_h, REML = FALSE)) %>%
    filter(effect == "fixed") %>%
    mutate(term = str_replace(term, "cluster_label", "")) %>%
    mutate(term = recode(term, `scale(Microglia)` = "Microglia", `scale(depth_from_pia_um)` = "Abs. depth")) 
  tidy_df$ephys_prop = f
  tidy_df$species = 'human'
  keep_df = bind_rows(keep_df, tidy_df)
  
  formula = paste0('scale(', f, ') ~ 1 + (1|corresponding_AIT2.3.1_alias) + Microglia + (1|donor_id)')
  print(formula)
  tidy_df = tidy(lmer(formula, data = metadata_m, REML = FALSE)) %>%
    filter(effect == "fixed") %>%
    mutate(term = str_replace(term, "corresponding_AIT2.3.1_alias", "")) %>%
    mutate(term = recode(term, `scale(Microglia)` = "Microglia", `scale(depth_from_pia_um)` = "Abs. depth")) 
  tidy_df$ephys_prop = f
  tidy_df$species = 'mouse'
  keep_df = bind_rows(keep_df, tidy_df)
}

keep_df$ephys_prop = factor(keep_df$ephys_prop, 
                                  levels = c("input_resistance","rheobase_i","v_baseline","sag","threshold_v", "trough_v", "apamp", "ahpamp", "width", "fi_fit_slope", "adapt"),
                                  labels = c("Input resistance","Rheobase","RMP","Sag ratio","AP thresh", "AP trough", "AP amp", "AHP amp", "AP width", "FI slope", "Adaptation"))
remove_ephys_props = c('AP amp', 'AHP amp')

# write a spreadsheet with the output of the stats modeling, including pvals and effect sizes
effects_df_output = keep_df %>% filter(term == 'Microglia', !ephys_prop %in% remove_ephys_props) %>%
  select(ephys_prop, species, estimate:p.value) %>% tbl_df()
write_csv(effects_df_output, file = './output/aibs_ephys_assocs.csv')

effects_df_output$sig = F
effects_df_output$sig[effects_df_output$p.value < 0.05] = T

effects_df_output = effects_df_output %>%
  mutate(conf.low = estimate - std.error * 1.96, 
         conf.high = estimate + std.error * 1.96, 
         y_asterisk_pos = case_when(sign(estimate) > 0 ~ conf.high + 0.1,
                                    sign(estimate) < 0 ~ conf.low - 0.1, 
                                    TRUE ~ 0))

bar_graph_plot = effects_df_output  %>% 
  ggplot(aes(x = ephys_prop, y = estimate, fill = species)) + 
  geom_bar(stat = "identity", position = "dodge2") +
  geom_errorbar(aes(x = ephys_prop, 
                    y = estimate, ymin = estimate - 1.96 * std.error, 
                    ymax = estimate +  1.96 * std.error), 
                position = "dodge2") +
  theme_classic() + 
  geom_text(aes(label = ifelse(sig, "*", ""), y = y_asterisk_pos, group = species), color = "black",
            position = position_dodge(width = .9), size = 20 / .pt) + 
  scale_fill_manual(values = c('#9632B8FF', '#EEA236FF')) +
  geom_hline(yintercept = 0) + 
  geom_text(aes(label = ifelse(sig, "*", ""), y = y_asterisk_pos, group = species), color = "black", position = position_dodge(width = .9), size = 20 / .pt) + 
  scale_fill_manual(values = c('#9632B8FF', '#EEA236FF', "dark green", 'light green')) +
  ylab('Microglial assoc (Beta)') + 
  theme_classic() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1, vjust = 1, size = 15), 
        axis.title.x = element_blank(),
        axis.text = element_text(size = 12),
        axis.title = element_text(size = 16), 
        legend.title=element_text(size=16),
        legend.text=element_text(size=12)
        ) 

bar_graph_plot
ggsave(filename = "./output/figure_4b.svg", device = "svg", width = 9, height = 6, units = "in", dpi = 500, limitsize = F)
ggsave(filename = "./output/figure_4b.png", device = "png", width = 9, height = 6, units = "in", dpi = 500, limitsize = F)

allen_institute_ephys_effects = effects_df_output

full_plot = plot_grid(ephys_example_plots, bar_graph_plot, nrow = 2, rel_heights = c(1.8, 1))
full_plot

```

### figure out which cells to use as examples for plotting ephys features for
```{r}
metadata_h %>% filter(cluster_label == 'FREM3', !is.na(depth_from_pia_um)) %>% group_by(donor_id) %>% 
  summarize(num_cells = n(), max_micro_score = max(Microglia), min_micro_score = min(Microglia), diff_micro_score = max_micro_score - min_micro_score) %>%   arrange(-diff_micro_score)
metadata_h %>% filter(donor_id == 'H18.03.314', cluster_label == 'FREM3') %>% arrange(depth_from_pia_um) %>% select(Microglia, cell_soma_normalized_depth, depth_from_pia_um, input_resistance, trough_v, everything()) %>% as.data.frame() %>% mutate(input_resistance = 10^input_resistance)
metadata_h %>% filter(donor_id == 'H18.06.368', cluster_label == 'FREM3') %>% arrange(cell_soma_normalized_depth) %>% select(Microglia, cell_soma_normalized_depth, depth_from_pia_um, input_resistance, trough_v, everything()) %>% as.data.frame() %>% mutate(input_resistance = 10^input_resistance)
metadata_h %>% filter(donor_id == 'H18.03.319', cluster_label == 'FREM3') %>% arrange(cell_soma_normalized_depth) %>% select(Microglia, cell_soma_normalized_depth, depth_from_pia_um, input_resistance, trough_v, everything()) %>% as.data.frame() %>% mutate(input_resistance = 10^input_resistance)

metadata_m %>%  group_by(donor_id, corresponding_AIT2.3.1_alias) %>% 
  summarize(num_cells = n(), max_micro_score = max(Microglia), min_micro_score = min(Microglia), diff_micro_score = max_micro_score - min_micro_score) %>%   arrange(-diff_micro_score)

metadata_m %>% filter(donor_id == '460700', corresponding_AIT2.3.1_alias == 'Sst Calb2 Pdlim5') %>% select(Microglia, cell_soma_normalized_depth, input_resistance, trough_v, everything()) %>% as.data.frame() %>% mutate(input_resistance = 10^input_resistance)
# low microglia cell example: 732048771	SM-GE8YK_S028_E1-50	H18.03.314	ephys_session_id sample_id donor_id 
# high microglia cell example: 732091715	SM-GE8YK_S031_E1-50	H18.03.314 ephys_session_id sample_id donor_id 
# maybe this pair too?
# low microglia cell example: 770369639	SM-GE5U8_S069_E1-50	H18.03.319	ephys_session_id sample_id donor_id
# high microglia cell example: 770466547	SM-GE5U8_S070_E1-50	H18.03.319 ephys_session_id sample_id donor_id


# do same as above for cells from gouwens
metadata_m %>% filter(cluster_label == 'Pvalb') %>% group_by(donor_id) %>% 
  summarize(num_cells = n(), max_micro_score = max(Microglia), min_micro_score = min(Microglia), diff_micro_score = max_micro_score - min_micro_score) %>%   arrange(-diff_micro_score)
```
